disablePushState typing
