+++ packages/nuxt-meta-pixel/src/typings.ts
@@ -1,0 +1,9 @@
+export interface Pixel {
+  id: string
+  noscript?: boolean
+  autoconfig?: boolean
+}
+
+export interface ModuleOptions {
+  pixels: Pixel[]
+}
